Even more partial changes to "RenameMePlease.sql"...

+++ ERCentralAPI/SQL Changes Manager/SQLScripts/RenameMePlease.sql
@@ -34,7 +34,9 @@ IF EXISTS(SELECT * FROM INFORMATION_SCHEMA.TABLES
 	 EXEC('INSERT INTO dbo.Tmp_tb_ITEM_MAG_MATCH (ITEM_ID, REVIEW_ID, PaperId, AutoMatchScore, ManualTrueMatch, ManualFalseMatch)
 		SELECT ITEM_ID, REVIEW_ID, PaperId, AutoMatchScore, ManualTrueMatch, ManualFalseMatch FROM dbo.tb_ITEM_MAG_MATCH WITH (HOLDLOCK TABLOCKX)')
 GO
-DROP TABLE dbo.tb_ITEM_MAG_MATCH
+IF EXISTS(SELECT * FROM INFORMATION_SCHEMA.TABLES
+           WHERE TABLE_NAME = N'tb_ITEM_MAG_MATCH')
+		   DROP TABLE dbo.tb_ITEM_MAG_MATCH
 GO
 EXECUTE sp_rename N'dbo.Tmp_tb_ITEM_MAG_MATCH', N'tb_ITEM_MAG_MATCH', 'OBJECT' 
 GO
@@ -95,7 +97,9 @@ CREATE NONCLUSTERED INDEX [NonClusteredIndex-20190911-093715] ON [dbo].[tb_ITEM_
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, SORT_IN_TEMPDB = OFF, DROP_EXISTING = OFF, ONLINE = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON)
 
 GO
-
+IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'[dbo].[st_ItemListMagNoMatches]') AND type in (N'P', N'PC'))
+DROP PROCEDURE [dbo].[st_ItemListMagNoMatches]
+GO
 
 /****** Object:  StoredProcedure [dbo].[st_ItemListMagNoMatches]    Script Date: 23/08/2019 22:59:36 ******/
 SET ANSI_NULLS ON
@@ -174,6 +178,8 @@ GO
 
 USE [Reviewer]
 GO
+
+
 /****** Object:  StoredProcedure [dbo].[st_ReviewContactForSiteAdmin]    Script Date: 08/09/2019 23:12:09 ******/
 SET ANSI_NULLS ON
 GO
@@ -256,6 +262,9 @@ GO
 
 USE [Reviewer]
 GO
+IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'[dbo].[st_CheckReviewHasUpdates]') AND type in (N'P', N'PC'))
+DROP PROCEDURE [dbo].[st_CheckReviewHasUpdates]
+GO
 /****** Object:  StoredProcedure [dbo].[st_CheckReviewHasUpdates]    Script Date: 21/09/2019 18:23:05 ******/
 SET ANSI_NULLS ON
 GO
@@ -266,7 +275,7 @@ GO
 -- Create date: 
 -- Description:	Checks to see whether a review has any auto-identified studies for authors to check
 -- =============================================
-ALTER PROCEDURE [dbo].[st_CheckReviewHasUpdates] 
+CREATE PROCEDURE [dbo].[st_CheckReviewHasUpdates] 
 	-- Add the parameters for the stored procedure here
 	@REVIEW_id int = 0,
 	@NUpdates INT OUTPUT
@@ -372,6 +381,9 @@ SET NOCOUNT OFF
 go
 
 USE [Reviewer]
+GO
+IF  EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'[dbo].[st_ItemListMaybeMagMatches]') AND type in (N'P', N'PC'))
+DROP PROCEDURE [dbo].[st_ItemListMaybeMagMatches]
 GO
 /****** Object:  StoredProcedure [dbo].[st_ItemListMaybeMagMatches]    Script Date: 18/08/2019 09:02:07 ******/
 SET ANSI_NULLS ON
